Title: 앞의 두 수를 더한 수열 2, Time: 3087ms, Memory: 8MB, Status: Time Limit Exceed - CodeTree

+++ 240502/앞의 두 수를 더한 수열 2/a-sequence-by-adding-the-previous-two-numbers-2.java
@@ -6,19 +6,19 @@ public class Main {
     static String testcase = "11";
 
     static int n;
-    static int[] memo;
+    static long[] memo;
 
     public static void main(String[] args) throws Exception {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 //        BufferedReader br = new BufferedReader(new StringReader(testcase));
 
         n = Integer.parseInt(br.readLine());
-        memo = new int[n];
+        memo = new long[n];
 
         System.out.println(fibo(n-1));
     }
 
-    static int fibo(int depth){
+    static long fibo(int depth){
         if(depth == 0 || depth == 1) {
             memo[depth] = 1;
             return memo[depth];
